requirements.txt
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
conda==4.5.11
cryptography==2.3.1
idna==2.7
Pillow==5.4.1
pycosat==0.6.3
pycparser==2.18
pyOpenSSL==18.0.0
PyQRCode==1.2.1
PySocks==1.6.8
qrcode==6.1
requests==2.20.0
ruamel-yaml==0.15.46
six==1.11.0
urllib3==1.23
